[grid] Moving to Docker API 1.41
