+++ java/server/src/org/openqa/selenium/docker/VersionCommand.java
@@ -17,12 +17,9 @@
 
 package org.openqa.selenium.docker;
 
-import static org.openqa.selenium.json.Json.MAP_TYPE;
-import static org.openqa.selenium.remote.http.HttpMethod.GET;
-
 import com.google.common.collect.ImmutableMap;
 
-import org.openqa.selenium.docker.v1_40.V140Docker;
+import org.openqa.selenium.docker.v1_41.V141Docker;
 import org.openqa.selenium.internal.Require;
 import org.openqa.selenium.json.Json;
 import org.openqa.selenium.json.JsonException;
@@ -36,12 +33,15 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.function.Function;
 
+import static org.openqa.selenium.json.Json.MAP_TYPE;
+import static org.openqa.selenium.remote.http.HttpMethod.GET;
+
 class VersionCommand {
 
   private static final Json JSON = new Json();
   // Insertion order matters, and is preserved by ImmutableMap.
-  private static final Map<Version, Function<HttpHandler, DockerProtocol>> SUPPORTED_VERSIONS = ImmutableMap.of(
-    new Version("1.40"), V140Docker::new);
+  private static final Map<Version, Function<HttpHandler, DockerProtocol>>
+    SUPPORTED_VERSIONS = ImmutableMap.of(new Version("1.40"), V141Docker::new);
 
   private final HttpHandler handler;
