+++ java/server/src/org/openqa/selenium/docker/v1_41/IsContainerPresent.java
@@ -15,45 +15,32 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.openqa.selenium.docker.v1_40;
-
-import static java.net.HttpURLConnection.HTTP_OK;
-import static org.openqa.selenium.remote.http.HttpMethod.GET;
+package org.openqa.selenium.docker.v1_41;
 
 import org.openqa.selenium.docker.ContainerId;
-import org.openqa.selenium.docker.ContainerLogs;
 import org.openqa.selenium.internal.Require;
-import org.openqa.selenium.remote.http.Contents;
 import org.openqa.selenium.remote.http.HttpHandler;
 import org.openqa.selenium.remote.http.HttpRequest;
 import org.openqa.selenium.remote.http.HttpResponse;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.logging.Logger;
+import static org.openqa.selenium.docker.v1_41.V141Docker.DOCKER_API_VERSION;
+import static org.openqa.selenium.remote.http.HttpMethod.GET;
 
-class GetContainerLogs {
-  private static final Logger LOG = Logger.getLogger(GetContainerLogs.class.getName());
+class IsContainerPresent {
   private final HttpHandler client;
 
-  public GetContainerLogs(HttpHandler client) {
-    this.client = Require.nonNull("HTTP client", client);
+  public IsContainerPresent(HttpHandler client) {
+    this.client = Require.nonNull("Http client", client);
   }
 
-  public ContainerLogs apply(ContainerId id) {
+  public boolean apply(ContainerId id) {
     Require.nonNull("Container id", id);
 
-    String requestUrl =
-      String.format("/v1.40/containers/%s/logs?stdout=true&stderr=true", id);
-
     HttpResponse res = client.execute(
-      new HttpRequest(GET, requestUrl)
+      new HttpRequest(GET, String.format("/v%s/containers/%s/json", DOCKER_API_VERSION, id))
         .addHeader("Content-Length", "0")
         .addHeader("Content-Type", "text/plain"));
-    if (res.getStatus() != HTTP_OK) {
-      LOG.warning("Unable to inspect container " + id);
-    }
-    List<String> logLines = Arrays.asList(Contents.string(res).split("\n"));
-    return new ContainerLogs(id, logLines);
+
+    return res.isSuccessful();
   }
 }
